Refactor ISM managed indexes classes
